popUpView refactoring

+++ ShoppingList/Binding/NewListBinding.swift
@@ -2,7 +2,7 @@ import UIKit
 
 enum NewListBinding {
     case updateCompleteButtonState
-    case showPopUp(Int, Int, Units)
+    case showPopUp(Int, Float, Units)
     case updateItem(IndexPath, Bool)
     case insertItem(IndexPath)
     case removeItem(IndexPath)
