+++ ShoppingList/Scenes/PopUpView/Protocols/PopUpVCDelegate.swift
@@ -1,4 +1,4 @@
 protocol PopUpVCDelegate: AnyObject {
     func unitSelected(item: Int, unit: Units)
-    func quantitySelected(item: Int, quantity: Int)
+    func quantitySelected(item: Int, quantity: Float)
 }
